ire_fold.h: quality score uses float

+++ src/ire_fold.h
@@ -6,7 +6,7 @@ typedef unsigned int uint;
 typedef struct IreStructure {
 	const char *sequence;    /* Sequence of the associated IRE */
 	char       *structure;   /* Predicted structure of the IRE */
-	uint        quality;     /* Quality score of the predicted structure */
+	float       quality;     /* Quality score of the predicted structure */
 
 	char       *__struct;    /* Stores structure that is being processed */
 	uint        __bestscore; /* Current best score in the preediction */
